Implemented 20x20 pixels, choosing correct face
This includes a new function where the roi in detectFaces_V2.m chooses the biggest face (tested and it seemed to work), and the code for separating the detected face into 20x20 pixels is made, but it is not runnable yet. I still need to sort out the top regions based on intensity for ICA to handel, then the main parts of the MRC algorithm will be implemented.