renamed files
Because we are trying a bunch of different methods for analyzing the face data, I think that it will reduce confusion just to keep different versions of the MRC algorithm file so that we know that the latest version is. I didn’t think that we would have so many versions of the code, but it is apparent that we need to keep them as we finalize which methods of face detection are the most accurate.

There were also some duplicate files in different folders. For instance, detectfaces_V2.m was in two different folders, so now it is in only the “faces” folder. I deleted the other copy.

I have renamed the versions of the while loop code to the following: “faces_MRC_WhileLoopTester.m” is now called “facesMRC_V1.m” and has the old ICA that we are not using anymore; “facesMRC_V2” has an old cheek and forehead detector, and “facesMRC_V3” is the code with the latest cheek and forehead detector that I am working on right now.

The imageRecorder.m file doesn’t really fit into any of the files, so it is outside of all three files in this branch. The function files are named as they were before.

I apologize for the confusion. Hopefully with the facesMRC files renamed this way, it will be easier to know what the latest version of the code is. I will keep working on the new cheek and face detection, and then I will code in the Goodness metric from the MRC algorithm to see if it makes a difference in the code. These updates will all be in facesMRC_V3.m.