fixed syntax and logic erros
The code runs! It seems to comtinuously collect images and analze them. The program still prints a lot of intermediate values and variables into the console for debugging, and I have kept them in so that when we implement the face-grid part of the MRC algorithm, we can have the debugging prit statements for then.